Completely rethought the WindowService/NavigationBar design
The need to be able to put in the navigation bar, an image which is not necessarily associated to a file lead to identify the group of window with something else than a QString.
That's why I introduced the concept of Node to represent groups of
window.
Nodes are now identified by the structure NodeId to make abstraction of the
underlying type.