Moved the QGraphicView from ImageView to ImageWindow, added the possibility to zoom outside the ImageView (inside the ImageWindow), optimized the ImageWindow Rendering.
The ImageView inherits QGraphicsPixmapItem, therefore as an item of a
QGraphicScene, it cannot create en destroy the QGraphicsScene and the
QGraphicsView. It's not only illogical, it's not technically possible
since the QGraphicsScene destroys all items it contains when it is
destroyed. It is much more logical and useful to have the ImageWindow
take care of the Scene/View, it also offer the possibility to use the
ImageView as any other QGraphicsItem by adding it to a layout or else.