Completely rethought the StandardImageView
StandardImageView is not a QGraphicsPixmapItem anymore, there were no
need for the QGraphics functionalities.
It's now a QScrollArea containing aQWidget on which the image is
directly paint as a QPixmap.
This is much more lightweight and allow the possibility to use the
opengl graphic acceleration if needed.
The state mode (_move, _resize) has also been rethought, the
StandardImageView work on two state modes, some buttons was added to the
StandardImageWindow.